+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [6.0.0] - Unreleased
+## [6.0.0] - 2020-08-25
 
 ### Fixed
 
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New algorithm to pretty print which fits more on a line if possible
 - Deprecated `character` parameter in Rule and Console.rule, in favor of `characters`
+- Optimized Syntax.from_path to avoid searching all lexers, which also speeds up tracebacks
 
 ### Added